requirements.txt
torch
pytorch-lightning
transformers
sentence_transformers
datasets
toml
gradio
lycoris-lora>=2.1.0.dev1
hakuphi>=0.0.1.dev1
prodigy
